backend/requirements.txt
Flask
markupsafe
werkzeug
redis==4.3.4
mysql-connector-python==8.0.28

